requirements.txt
streamlit==1.36.0
openai==1.35.15
requests==2.32.3
numpy==2.0.0
scipy==1.13.1
streamlit-audiorecorder==0.0.5